Menus, saves, better shadows

+++ OpenGLTest/TweenAction.h
@@ -46,18 +46,18 @@ template <typename T>
 T TweenAction<T>::EvalAt(float time)
 {
 	float t = ((time - startTime) / duration);
-	double amount;
+	float amount;
 
 	switch (easeType)
 	{
 	case EaseType::sineIn:
-		amount = 1 - cos((t * PI) / 2);
+		amount = 1.0f - (float)cos((t * PI) / 2);
 		break;
 	case EaseType::sineOut:
-		amount = sin((t * PI) / 2);
+		amount = (float)sin((t * PI) / 2);
 		break;
 	case EaseType::sineInOut:
-		amount = -(cos(PI * t) - 1) / 2;
+		amount = -((float)cos(PI * t) - 1) / 2;
 		break;
 	case EaseType::quadIn:
 		amount = t * t;
@@ -66,52 +66,52 @@ T TweenAction<T>::EvalAt(float time)
 		amount = 1 - (1 - t) * (1 - t);
 		break;
 	case EaseType::quadInOut:
-		amount = t < 0.5 ? 2 * t * t : 1 - pow(-2 * t + 2, 2) / 2;
+		amount = t < 0.5f ? 2.0f * t * t : 1 - (float)pow(-2 * t + 2, 2) / 2;
 		break;
 	case EaseType::cubicIn:
 		amount = t * t * t;
 		break;
 	case EaseType::cubicOut:
-		amount = 1 - pow(1 - t, 3);
+		amount = 1 - (float)pow(1 - t, 3);
 		break;
 	case EaseType::cubicInOut:
-		amount = t < 0.5 ? 4 * t * t * t : 1 - pow(-2 * t + 2, 3) / 2;
+		amount = t < 0.5f ? 4.0f * t * t * t : 1 - (float)pow(-2 * t + 2, 3) / 2;
 		break;
 	case EaseType::quartIn:
 		amount = t * t * t * t;
 		break;
 	case EaseType::quartOut:
-		amount = 1 - pow(1 - t, 4);
+		amount = 1 - (float)pow(1 - t, 4);
 		break;
 	case EaseType::quartInOut:
-		amount = t < 0.5 ? 8 * pow(t, 4) : 1 - pow(-2 * t + 2, 4) / 2;
+		amount = t < 0.5f ? 8 * (float)pow(t, 4) : 1 - (float)pow(-2 * t + 2, 4) / 2;
 		break;
 	case EaseType::quintIn:
-		amount = pow(t, 5);
+		amount = (float)pow(t, 5);
 		break;
 	case EaseType::quintOut:
-		amount = 1 - pow(1 - t, 5);
+		amount = 1 - (float)pow(1 - t, 5);
 		break;
 	case EaseType::quintInOut:
-		amount = t < 0.5 ? 16 * pow(t, 5) : 1 - pow(-2 * t + 2, 5) / 2;
+		amount = t < 0.5f ? 16 * (float)pow(t, 5) : 1 - (float)pow(-2 * t + 2, 5) / 2;
 		break;
 	case EaseType::expoIn:
-		amount = t == 0 ? 0 : pow(2, 10 * t - 10);
+		amount = t == 0 ? 0 : (float)pow(2, 10 * t - 10);
 		break;
 	case EaseType::expoOut:
-		amount = t == 1 ? 1 : 1 - pow(2, -10 * t);
+		amount = t == 1 ? 1 : 1 - (float)pow(2, -10 * t);
 		break;
 	case EaseType::expoInOut:
-		amount = t == 0 ? 0 : t == 1 ? 1 : t < 0.5 ? pow(2, 20 * t - 10) / 2 : (2 - pow(2, -20 * t + 10)) / 2;
+		amount = t == 0 ? 0 : t == 1 ? 1 : t < 0.5 ? (float)pow(2, 20 * t - 10) / 2 : (2 - (float)pow(2, -20 * t + 10)) / 2;
 		break;
 	case EaseType::circIn:
-		amount = 1 - sqrt(1 - pow(t, 2));
+		amount = 1 - (float)sqrt(1 - (float)pow(t, 2));
 		break;
 	case EaseType::circOut:
-		amount = sqrt(1 - pow(t - 1, 2));
+		amount = (float)sqrt(1 - (float)pow(t - 1, 2));
 		break;
 	case EaseType::circInOut:
-		amount = t < 0.5 ? (1 - sqrt(1 - pow(2 * t, 2))) / 2 : (sqrt(1 - pow(-2 * t + 2, 2)) + 1) / 2;
+		amount = t < 0.5f ? (1 - (float)sqrt(1 - (float)pow(2 * t, 2))) / 2 : ((float)sqrt(1 - (float)pow(-2 * t + 2, 2)) + 1) / 2;
 		break;
 	default:
 		amount = t;
